# public/swagger.yaml
openapi: 3.0.3
info:
  title: Food Delivery System API
  description: |
    RESTful API for Food Delivery System with RAG Chatbot

    ## Features
    - User authentication & authorization
    - Restaurant & menu management
    - Order processing & tracking
    - Real-time delivery tracking
    - Push notifications
    - Chat messaging
    - RAG-powered chatbot

    ## Base URL
    - Development: `http://localhost:3000`
  version: 1.0.0
  contact:
    name: API Support
    email: support@fooddelivery.com
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/restaurants:
    get:
      tags:
        - Restaurants
      summary: Get all restaurants
      parameters:
        - name: isActive
          in: query
          schema:
            type: boolean
        - name: isOpen
          in: query
          schema:
            type: boolean
        - name: ownerId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of restaurants
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Restaurant'
  /api/restaurants/{id}:
    get:
      tags:
        - Restaurants
      summary: Get restaurant by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Restaurant details
...
  /api/orders:
    get:
      tags:
        - Orders
      summary: Get orders
      parameters:
        - name: customerId
          in: query
          schema:
            type: string
        - name: restaurantId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of orders

    post:
      tags:
        - Orders
      summary: Create new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrder'
      responses:
        '201':
          description: Order created successfully
...
components:
  schemas:
    Restaurant:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        rating:
          type: number
        deliveryFee:
          type: number
        minimumOrder:
          type: number
        estimatedTime:
          type: string
        isOpen:
          type: boolean
...
    CreateOrder:
      type: object
      required:
        - customerId
        - restaurantId
        - addressId
        - items
        - subtotal
        - deliveryFee
        - total
        - paymentMethod
      properties:
        customerId:
          type: string
        restaurantId:
          type: string
        addressId:
          type: string
        items:
          type: array
          items:
            type: object
            properties:
              menuItemId:
                type: string
              quantity:
                type: integer
              price:
                type: number
        subtotal:
          type: number
        deliveryFee:
          type: number
        total:
          type: number
        paymentMethod:
          type: string
          enum: [CASH, CREDIT_CARD, DEBIT_CARD, MOBILE_BANKING, PROMPTPAY]
        notes:
          type: string